requirements.txt
folium==0.18.0
haversine==2.8.1
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
plotly==5.9.0
streamlit==1.39.0
streamlit_folium==0.23.1
